+++ src/Dashboard/Drawings/useCreateDrawing.ts
@@ -1,6 +1,6 @@
 import { collection, serverTimestamp, addDoc } from 'firebase/firestore';
 
-import { db } from 'shared/services/firebase';
+import { db } from 'shared/utils/firebase';
 import useCurrentUser from 'shared/hooks/useCurrentUser';
 
 const useCreateDrawing = () => {
