+++ src/Dashboard/Drawings/useDrawings.ts
@@ -8,7 +8,7 @@ import {
   DocumentData,
 } from 'firebase/firestore';
 
-import { db } from 'shared/services/firebase';
+import { db } from 'shared/utils/firebase';
 import useCurrentUser from 'shared/hooks/useCurrentUser';
 
 const useDrawings = () => {
