+++ src/Dashboard/useSignOut.ts
@@ -1,6 +1,6 @@
 import { signOut as signOutUser } from 'firebase/auth';
 
-import { auth } from 'shared/services/firebase';
+import { auth } from 'shared/utils/firebase';
 
 const useSignOut = () => {
   const signOut = () => signOutUser(auth);
