+++ src/shared/components/Input/Input.tsx
@@ -1,6 +1,6 @@
 import React, { FC, InputHTMLAttributes, ChangeEventHandler } from 'react';
 
-import icons from 'shared/services/icons';
+import icons from 'shared/utils/icons';
 
 import Icon from 'shared/components/Icon/Icon';
