+++ src/shared/utils/firebase.ts
